+++ src/header/Stash Markers Autoscroll.user.js
@@ -1,0 +1,11 @@
+// ==UserScript==
+// @name        Stash Markers Autoscroll
+// @namespace   %NAMESPACE%
+// @description Automatically scrolls markers page
+// @version     0.1.0
+// @author      7dJx1qP
+// @match       %MATCHURL%
+// @grant       unsafeWindow
+// @require     %LIBRARYPATH%
+// @require     %FILEPATH%
+// ==/UserScript==